backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
httpx==0.28.1
python-dotenv==1.0.1
openai==1.57.2
pdfplumber==0.11.4
python-docx==1.1.2
